STARTING SERVICE ONLY IN PRODUCTION ENVIRONMENT

+++ src/background.js
@@ -33,28 +33,21 @@ if (env.name !== 'production') {
 }
 
 app.on('ready', function () {
-    const jarPath = __dirname + '\\..\\service\\backend.jar';
-    const exec = require('child_process').exec;
-    /*const childProcess = exec('start "gnos_service" java -jar ' + jarPath,
-        (error, stdout, stderr) => {
-            console.log('stdout: ' + stdout);
-            console.log('stderr: ' + stderr);
-            if (error !== null) {
-                console.log('exec error: ' + error);
+    if (env.name === 'production') {
+        const jarPath = __dirname + '\\..\\service\\backend.jar';
+        const exec = require('child_process').exec;
+        const childProcess = exec('javaw -jar ' + jarPath,
+            (error, stdout, stderr) => {
+                console.log('stdout: ' + stdout);
+                console.log('stderr: ' + stderr);
+                if (error !== null) {
+                    console.log('exec error: ' + error);
+                }
             }
-        }
-     )*/
-    const childProcess = exec('javaw -jar ' + jarPath,
-        (error, stdout, stderr) => {
-            console.log('stdout: ' + stdout);
-            console.log('stderr: ' + stderr);
-            if (error !== null) {
-                console.log('exec error: ' + error);
-            }
-        }
-    )
-    childProcessId = childProcess.pid;
-    console.log('Starting pid: ' + childProcessId);
+        )
+        childProcessId = childProcess.pid;
+        console.log('Starting pid: ' + childProcessId);
+    }
     setApplicationMenu();
 
     var mainWindow = createWindow('main', {
@@ -72,25 +65,18 @@ app.on('ready', function () {
 });
 
 app.on('window-all-closed', function () {
-    const exec = require('child_process').exec;
-    /*const child = exec('TASKKILL /FI "WINDOWTITLE eq gnos_service"',
-        (error, stdout, stderr) => {
-            console.log('stdout: ' + stdout);
-            console.log('stderr: ' + stderr);
-            if (error !== null) {
-                console.log('exec error: ' + error);
+    if (env.name === 'production') {
+        const exec = require('child_process').exec;
+        console.log('Closing pid: ' + childProcessId);
+        childProcess = exec('taskkill /pid ' + childProcessId,
+            (error, stdout, stderr) => {
+                console.log('stdout: ' + stdout);
+                console.log('stderr: ' + stderr);
+                if (error !== null) {
+                    console.log('exec error: ' + error);
+                }
             }
-        }
-     )*/
-    console.log('Closing pid: ' + childProcessId);
-    childProcess = exec('taskkill /pid ' + childProcessId,
-        (error, stdout, stderr) => {
-            console.log('stdout: ' + stdout);
-            console.log('stderr: ' + stderr);
-            if (error !== null) {
-                console.log('exec error: ' + error);
-            }
-        }
-    )
+        )
+    }
     app.quit();
 });
